# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "contact": {
      "name": "NIEMOpen",
      "url": "https://niemopen.org"
    },
    "description": "Community support for common NIEM tool functionality to browse and search the model, build and migrate subsets, and check conformance",
    "title": "NIEM API 2.0",
    "version": "2.0-alpha07:SNAPSHOT"
  },
  "servers": [
    {
      "url": "https://tools.niem.gov/api/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080/api/v2",
      "description": "Local machine"
    },
    {
      "url": "https://yellowkiwi06.icl.gtri.org/",
      "description": "Internal test server"
    }
  ],
...
  "paths": {
...
    "/validation/cmf/xml": {
      "post": {
        "description": "Validate a CMF XML file against the NIEM CMF XML schema.",
        "operationId": "getCmfXMLValidation",
        "parameters": [
          {
            "description": "REQUEST BODY PARAMETER. File format for the validation results.",
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "A CMF XML file."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "text/csv": {}
            },
            "description": "Success"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "500": {
            "description": "Error"
          }
        },
        "summary": "Validate a CMF XML file against the NIEM CMF XML schema.",
        "tags": [
          "Validation"
        ]
      }
    },
    "/validation/instances/xml": {
      "post": {
        "description": "Validate one or more XML instances against XML schemas.",
        "operationId": "getXMLValidation",
        "parameters": [
          {
            "description": "REQUEST BODY PARAMETER. File format for the validation results.",
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "xml": {
                    "type": "string",
                    "format": "binary",
                    "description": "An XML instance file or set of XML instance files in a zip file."
                  },
                  "xsd": {
                    "type": "string",
                    "format": "binary",
                    "description": "An XML schema or set of XML schemas in a zip file."
                  }
                },
                "required": [
                  "xml",
                  "xsd"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "text/csv": {}
            },
            "description": "Success"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "500": {
            "description": "Error"
          }
        },
        "summary": "Validate one or more XML instances against XML schemas.",
        "tags": [
          "Validation"
        ]
      }
    },
    "/validation/message-catalog": {
      "post": {
        "description": "Validate an IEPD / message catalog instance file against the NIEM IEPD /\n message catalog schema.",
        "operationId": "getMessageCatalogXMLValidation",
        "parameters": [
          {
            "description": "REQUEST BODY PARAMETER. File format for the validation results.",
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "An IEPD / message catalog instance file"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "text/csv": {}
            },
            "description": "Success"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "500": {
            "description": "Error"
          }
        },
        "summary": "Validate an IEPD / message catalog instance file against the NIEM IEPD /\n message catalog schema.",
        "tags": [
          "Validation"
        ]
      }
    },
    "/validation/message-specification": {
      "post": {
        "description": "Validate a message specification or IEPD zip file against NIEM message\n specification / IEPD conformance rules.\n\n Runs the following checks:\n\n - Validates the MPD or IEPD catalog.\n - Validates XML catalogs\n - Validates XML schemas\n - Validates XML schemas against NDR conformance rules",
        "operationId": "getMessageSpecificationValidation",
        "parameters": [
          {
            "description": "REQUEST BODY PARAMETER. File format for the validation results.",
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "A message specification or IEPD zip file."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "text/csv": {}
            },
            "description": "Success"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "500": {
            "description": "Error"
          }
        },
        "summary": "Validate a message specification or IEPD zip file against NIEM message\n specification / IEPD conformance rules.",
        "tags": [
          "Validation"
        ]
      }
    },
    "/validation/schemas/ndr": {
      "post": {
        "description": "Validate one or more XML Schemas against the NIEM Naming and Design Rules (NDR).",
        "operationId": "getSchemaNDRValidation",
        "parameters": [
          {
            "description": "REQUEST BODY PARAMETER. File format for the validation results.",
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "XML Schema or zip file"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "text/csv": {}
            },
            "description": "Success"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "500": {
            "description": "Error"
          }
        },
        "summary": "Validate one or more XML Schemas against the NIEM Naming and Design Rules (NDR).",
        "tags": [
          "Validation"
        ]
      }
    },
    "/validation/schemas/xml": {
      "post": {
        "description": "Validate one or more XML Schemas.",
        "operationId": "getXsdValidation",
        "parameters": [
          {
            "description": "File format for the validation results.",
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "REQUEST BODY PARAMETER.  Single XML Schema or a zip file containing\n multiple XML Schemas."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "text/csv": {}
            },
            "description": "Success"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "500": {
            "description": "Error"
          }
        },
        "summary": "Validate one or more XML Schemas.",
        "tags": [
          "Validation"
        ]
      }
    },
    "/validation/xml-catalog": {
      "post": {
        "description": "Validate a XML catalog against the OASIS catalog schema.",
        "operationId": "getXmlCatalogXmlValidation",
        "parameters": [
          {
            "description": "REQUEST BODY PARAMETER. File format for the validation results.",
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "An XML catalog file."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "text/csv": {}
            },
            "description": "Success"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "500": {
            "description": "Error"
          }
        },
        "summary": "Validate a XML catalog against the OASIS catalog schema.",
        "tags": [
          "Validation"
        ]
      }
    },
...
  },
...
}